refactor: parametrize dataset options
